backend/requirements.txt
python-dotenv
psycopg2
flask
flask-sqlalchemy
flask-cors
flask-caching